# docs/swagger.yaml
basePath: /
host: localhost:8080
definitions:
...
  recipebox-backend-go_internal_dto.DashboardEnvelope:
    properties:
      data:
        $ref: '#/definitions/recipebox-backend-go_internal_dto.DashboardResponse'
    type: object
  recipebox-backend-go_internal_dto.DashboardMealPlan:
    properties:
      day:
        type: string
      id:
        type: integer
      ingredients:
        items:
          type: string
        type: array
      mealName:
        type: string
      servings:
        type: integer
    type: object
  recipebox-backend-go_internal_dto.DashboardRecipe:
    properties:
      category:
        type: string
      id:
        type: integer
      name:
        type: string
      prepTime:
        type: integer
      updatedAt:
        type: string
    type: object
  recipebox-backend-go_internal_dto.DashboardResponse:
    properties:
      mealPlans:
        items:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.DashboardMealPlan'
        type: array
      recipes:
        items:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.DashboardRecipe'
        type: array
      shoppingItems:
        items:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.DashboardShoppingItem'
        type: array
      summary:
        $ref: '#/definitions/recipebox-backend-go_internal_dto.DashboardSummary'
    type: object
  recipebox-backend-go_internal_dto.DashboardShoppingItem:
    properties:
      checked:
        type: boolean
      id:
        type: integer
      menuName:
        type: string
      name:
        type: string
      qty:
        type: string
    type: object
  recipebox-backend-go_internal_dto.DashboardSummary:
    properties:
      pendingShoppingItemCount:
        type: integer
      recipeCount:
        type: integer
      upcomingMealPlanCount:
        type: integer
    type: object
...
  recipebox-backend-go_internal_dto.MealPlanEnvelope:
    properties:
      data:
        $ref: '#/definitions/recipebox-backend-go_internal_dto.DashboardMealPlan'
    type: object
  recipebox-backend-go_internal_dto.MealPlansEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.DashboardMealPlan'
        type: array
    type: object
  recipebox-backend-go_internal_dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  recipebox-backend-go_internal_dto.RecipeEnvelope:
    properties:
      data:
        $ref: '#/definitions/recipebox-backend-go_internal_dto.DashboardRecipe'
    type: object
  recipebox-backend-go_internal_dto.RecipesEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.DashboardRecipe'
        type: array
    type: object
...
  recipebox-backend-go_internal_dto.UpsertMealPlanRequest:
    properties:
      day:
        type: string
      ingredients:
        items:
          type: string
        type: array
      mealName:
        type: string
      servings:
        type: integer
    type: object
  recipebox-backend-go_internal_dto.UpsertRecipeRequest:
    properties:
      category:
        type: string
      name:
        type: string
      prepTime:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for RecipeBox backend services.
  title: RecipeBox Backend API
  version: "1.0"
paths:
...
  /api/v1/dashboard:
    get:
      description: Return recipes, meal plans, shopping items and summary in one payload.
      parameters:
      - description: 'Bearer access token. Format: Bearer <access_token>'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.DashboardEnvelope'
      summary: Get dashboard overview
      tags:
      - Dashboard
  /api/v1/meal-plans:
    get:
      description: Return meal plans list used by dashboard/meal-plan page.
      parameters:
      - description: 'Bearer access token. Format: Bearer <access_token>'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Max items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.MealPlansEnvelope'
      summary: List meal plans
      tags:
      - Meal Plans
    post:
      consumes:
      - application/json
      description: Create a meal plan menu item.
      parameters:
      - description: 'Bearer access token. Format: Bearer <access_token>'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Meal plan payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.UpsertMealPlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.MealPlanEnvelope'
      summary: Create meal plan
      tags:
      - Meal Plans
  /api/v1/meal-plans/{id}:
    delete:
      description: Delete a meal plan menu item by ID.
      parameters:
      - description: 'Bearer access token. Format: Bearer <access_token>'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Meal Plan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.MessageResponse'
      summary: Delete meal plan
      tags:
      - Meal Plans
    put:
      consumes:
      - application/json
      description: Update a meal plan menu item by ID.
      parameters:
      - description: 'Bearer access token. Format: Bearer <access_token>'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Meal Plan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Meal plan payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.UpsertMealPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.MealPlanEnvelope'
      summary: Update meal plan
      tags:
      - Meal Plans
  /api/v1/recipes:
    get:
      description: Return recipes list used by dashboard/recipes page.
      parameters:
      - description: 'Bearer access token. Format: Bearer <access_token>'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Max items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.RecipesEnvelope'
      summary: List recipe cards
      tags:
      - Recipes
    post:
      consumes:
      - application/json
      description: Create a recipe menu item.
      parameters:
      - description: 'Bearer access token. Format: Bearer <access_token>'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Recipe payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.UpsertRecipeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.RecipeEnvelope'
      summary: Create recipe
      tags:
      - Recipes
  /api/v1/recipes/{id}:
    delete:
      description: Delete a recipe menu item by ID.
      parameters:
      - description: 'Bearer access token. Format: Bearer <access_token>'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.MessageResponse'
      summary: Delete recipe
      tags:
      - Recipes
    put:
      consumes:
      - application/json
      description: Update a recipe menu item by ID.
      parameters:
      - description: 'Bearer access token. Format: Bearer <access_token>'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      - description: Recipe payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/recipebox-backend-go_internal_dto.UpsertRecipeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipebox-backend-go_internal_dto.RecipeEnvelope'
      summary: Update recipe
      tags:
      - Recipes
...
schemes:
- http
- https
swagger: "2.0"